[fix] 파일 경로 수정 #63
