--- booking/src/main/java/com/pyokemon/booking/repository/BookingRepository.java
@@ -1,4 +1,0 @@
-package com.pyokemon.booking.repository;
-
-public interface BookingRepository {
-}
