+++ booking/src/main/java/com/pyokemon/payment/BookingApplication.java
@@ -1,14 +1,14 @@
-package com.pyokemon.booking;
+package com.pyokemon.payment;
 
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication(scanBasePackages = {"com.pyokemon"})
-@MapperScan("com.pyokemon.payment.repository")
-public class PaymentApplication {
+@MapperScan("com.pyokemon.booking.repository")
+public class BookingApplication {
   public static void main(String[] args) {
 
-    SpringApplication.run(PaymentApplication.class, args);
+    SpringApplication.run(BookingApplication.class, args);
   }
 }
