+++ booking/src/main/java/com/pyokemon/payment/repository/BookingRepository.java
@@ -1,0 +1,4 @@
+package com.pyokemon.payment.repository;
+
+public interface BookingRepository {
+}
