+++ booking/src/main/resources/mapper/BookingMapper.xml
@@ -3,6 +3,6 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.pyokemon.booking.repository.BookingRepository">
+<mapper namespace="com.pyokemon.payment.repository.BookingRepository">
 
 </mapper>
