--- payment/src/main/java/com/pyokemon/booking/repository/PaymentRepository.java
@@ -1,4 +1,0 @@
-package com.pyokemon.booking.repository;
-
-public interface PaymentRepository {
-}
