+++ payment/src/main/java/com/pyokemon/payment/repository/PaymentRepository.java
@@ -1,0 +1,4 @@
+package com.pyokemon.payment.repository;
+
+public interface PaymentRepository {
+}
